+++ src/settings.js
@@ -1985,7 +1985,7 @@ var MIN_FIREFOX_VERSION = 79;
 // older, i.e. iPhone 4s, iPad 2, iPad 3, iPad Mini 1, Pod Touch 5 and older,
 // see https://github.com/emscripten-core/emscripten/pull/7191.
 // MAX_INT (0x7FFFFFFF, or -1) specifies that target is not supported.
-// Minimum supported value is 140100 which was released on 2021-04-26 (see
+// Minimum supported value is 150000 which was released on 2021-09-20 (see
 // feature_matrix.py).
 // [link]
 var MIN_SAFARI_VERSION = 150000;
@@ -2142,10 +2142,9 @@ var SEPARATE_DWARF_URL = '';
 // not in others like split-dwarf).
 // When this flag is turned on, we error at link time if the build requires any
 // changes to the wasm after link. This can be useful in testing, for example.
-// Some example of features that require post-link wasm changes are:
+// Some examples of features that require post-link wasm changes are:
 //
 // - Lowering i64 to i32 pairs at the JS boundary (See WASM_BIGINT)
-// - Lowering nontrapping-float-to-int operations when targeting older browsers.
 var ERROR_ON_WASM_CHANGES_AFTER_LINK = false;
 
 // Abort on unhandled exceptions that occur when calling exported WebAssembly
@@ -2221,8 +2220,8 @@ var TRUSTED_TYPES = false;
 // polyfills be included in the output.  If you would prefer to take care of
 // polyfilling yourself via some other mechanism you can prevent emscripten
 // from generating these by passing ``-sNO_POLYFILL`` or ``-sPOLYFILL=0``
-// With default browser targets emscripten does not need any polyfills so this
-// settings is *only* needed when also explicitly targeting older browsers.
+// Currently emscripten does not support targeting any browsers that require
+// polyfills so this setting does nothing right now.
 var POLYFILL = true;
 
 // If non-zero, add tracing to core runtime functions.  Can be set to 2 for