+++ tools/link.py
@@ -32,7 +32,6 @@ from . import (
   webassembly,
 )
 from .cmdline import OFormat, options
-from .feature_matrix import Feature
 from .minimal_runtime_shell import generate_minimal_runtime_html
 from .settings import (
   DEPRECATED_SETTINGS,
@@ -310,37 +309,8 @@ def should_run_binaryen_optimizer():
   return settings.OPT_LEVEL >= 2
 
 
-def get_binaryen_lowering_passes():
-  passes = []
-
-  # The following features are all enabled in llvm by default and therefore
-  # enabled in the emscripten system libraries.  This means that we need to
-  # lower them away using binaryen passes, if they are not enabled in the
-  # feature matrix.
-  # This can happen if the feature is explicitly disabled on the command line,
-  # or when targeting an VM/engine that does not support the feature.
-
-  # List of [<feature_name>, <lowering_flag>, <feature_flags>] triples.
-  features = [
-    [Feature.NON_TRAPPING_FPTOINT, '--llvm-nontrapping-fptoint-lowering', ['--enable-nontrapping-float-to-int']],
-    [Feature.BULK_MEMORY, '--llvm-memory-copy-fill-lowering', ['--enable-bulk-memory', '--enable-bulk-memory-opt']],
-  ]
-
-  for feature, lowering_flag, feature_flags in features:
-    if not feature_matrix.caniuse(feature):
-      logger.debug(f'lowering {feature.name} feature due to incompatible target browser engines')
-      for f in feature_flags:
-        # Remove features from binaryen_features, otherwise future runs of binaryen
-        # could re-introduce the feature.
-        if f in building.binaryen_features:
-          building.binaryen_features.remove(f)
-      passes.append(lowering_flag)
-
-  return passes
-
-
 def get_binaryen_passes():
-  passes = get_binaryen_lowering_passes()
+  passes = []
   optimizing = should_run_binaryen_optimizer()
 
   # safe heap must run before post-emscripten, so post-emscripten can apply the sbrk ptr
@@ -1098,7 +1068,6 @@ def phase_linker_setup(linker_args):  # ruff: ignore[complex-structure, too-many
     if user_settings.get('WASM_BIGINT') and settings.WASM_BIGINT:
       exit_with_error('WASM_BIGINT=1 is not compatible with wasm2js')
     settings.WASM_BIGINT = 0
-    feature_matrix.disable_feature(Feature.JS_BIGINT_INTEGRATION)
 
   if options.oformat == OFormat.WASM and not settings.SIDE_MODULE:
     # if the output is just a wasm file, it will normally be a standalone one,